refactor: Move schemafunc.AttributeChanges to customplanmodifier and remove unused CopyUnknowns in schemafunc
